+++ src/routes/game/[id=integer]/round/[ordinal=integer]/edit/_Cell.svelte
